Update package.json
Given that the version of optional dependencies is set with `^1.18.1`, any update to the MINOR number will cause the latest MAJOR version available of `@moonrepo/core-*` to be downloaded. This leads to the installation of Moon always downloading the binary of the latest version of version 1. For instance:

```
➜  cli npm list -g
/Users/david/.nvm/versions/node/v20.10.0/lib
├── @moonrepo/cli@1.17.4
├── corepack@0.22.0
└── npm@10.2.3
➜  cli pwd
/Users/david/.nvm/versions/node/v20.10.0/lib/node_modules/@moonrepo/cli
➜  cli ls moon
 moon
➜  cli rm -f moon 
➜  cli node postinstall.js
➜  cli ./moon --version
moon 1.18.1
➜  cli 
```

The expected result was that the version would be 1.17.4. "This is a bug that affects all versions ^1.x.y."