Feature for defining the update locations
In a cooperate environment most internet resources
are restriced to access, or company policy
describes that binaries shall have to be stored
in so called artifact products for security
scanning, like Artifacotry or Nexus.

With this feature implementation we are able
to define the locations for current information
json and the on-premise location within the
company.

Changelog: feat