fix: support git worktrees with bare repo
The current logic is broken for the worktree layout I was trying to
follow. It seemed a bit too brittle - always assuming there's going to
be a .git dir, if I understand correctly. This change shortcircuits the
worktree detection if it finds a .git file that has "gitdir:" in it,
which to my belief only happens if we're in a worktree - so we're good
to go on assuming we're in a worktree ;) Assuming that we'll always be
inside of a larger git repo that we can find the workspace route is
inaccurate.